full imp of expr handling and started with list record
Now specifications like N/6, 3, N*4 can appear in an
ENDF recipe specification and the value of the variable
is extracted from that expression and the corresponding
value (as calucled on the basis of the value in the ENDF
file stored in the dictionary.)
Also started with the implementation of the list record type.
Now the loops in a list must be implemented correctly.