modify MF8/MT457 recipe to avoid arrays with gaps
Missing indices can occur if the level of
detailed is different for distinct radiation energies.
Whereas the gaps don't pose a problem for the ENDF interpreter,
they can't be handled properly by the parser at present.
For the time being, I also consider it more user-friendly and
less confusing for the user to avoid arrays with missing indices.