start using get_numeric_field_getter and revise mode handling
Originally the idea was to create if-else statements to
perform either reading or writing based on mode. However,
the new approach is to define basic getter functions,
such as `get_numeric_field_getter` that depending on the
mode does different things, i.e. either creating cpp code
to read a value from a line or cpp code to read from a
Python dictionary.